I created objects that simulate the DNA of P. aequor for my research team to study.
Context: I am part of a fictional research team that has found a new mysterious organism at the bottom of the ocean near hydrothermal vents. My team names the organism, Pila aequor (P. aequor), and finds that it is only comprised of 15 DNA bases. The small DNA samples and frequency at which it mutates due to the hydrothermal vents make P. aequor an interesting specimen to study. However, P. aequor cannot survive above sea level and locating P. aequor in the deep sea is difficult and expensive.  My job is to create objects that simulate the DNA of P. aequor for my research team to study.